fix: move elapsed time to helpers
